+++ packages/engine/Specs/Core/EllipsoidSpec.js
@@ -825,6 +825,9 @@ describe("Core/Ellipsoid", function () {
     Ellipsoid.WGS84.radii.z,
   ]);
 
-  // TODO: ellipsoids default
-  // TODO: Ensure API defaults are adjusted, ie. atmopshere moon, etc...
+  it("set default throws if undefined", function () {
+    expect(function () {
+      Ellipsoid.default = undefined;
+    }).toThrowDeveloperError();
+  });
 });